requirements.txt
Django==5.2.8
oss2==2.18.6
python-dotenv==1.0.1
pymysql==1.1.1
python-dateutil==2.9.0.post0
